// Frontend/src/messages/ProtEntityType.d.ts
/* eslint-disable */
export enum ProtEntityType {
  PROT_ENTITY_TYPE_NONE = 0,
  PROT_ENTITY_TYPE_AVATAR = 1,
  PROT_ENTITY_TYPE_MONSTER = 2,
  PROT_ENTITY_TYPE_NPC = 3,
  PROT_ENTITY_TYPE_GADGET = 4,
  PROT_ENTITY_TYPE_REGION = 5,
  PROT_ENTITY_TYPE_WEAPON = 6,
  PROT_ENTITY_TYPE_WEATHER = 7,
  PROT_ENTITY_TYPE_SCENE = 8,
  PROT_ENTITY_TYPE_TEAM = 9,
  PROT_ENTITY_TYPE_MASSIVE_ENTITY = 10,
  PROT_ENTITY_TYPE_MP_LEVEL = 11,
  PROT_ENTITY_TYPE_PLAY_TEAM_ENTITY = 12,
  PROT_ENTITY_TYPE_EYE_POINT = 13,
  PROT_ENTITY_TYPE_MAX = 14,
  UNRECOGNIZED = -1,
}