// Frontend/src/messages/SceneGadgetInfo.d.ts
/* eslint-disable */
import type { GadgetBornType } from "./GadgetBornType.d.ts";
import type { PlatformInfo } from "./PlatformInfo.d.ts";
import type { GadgetPlayInfo } from "./GadgetPlayInfo.d.ts";
import type { Item } from "./Item.d.ts";
import type { GatherGadgetInfo } from "./GatherGadgetInfo.d.ts";
import type { WorktopInfo } from "./WorktopInfo.d.ts";
import type { ClientGadgetInfo } from "./ClientGadgetInfo.d.ts";
import type { WeatherInfo } from "./WeatherInfo.d.ts";
import type { AbilityGadgetInfo } from "./AbilityGadgetInfo.d.ts";
import type { StatueGadgetInfo } from "./StatueGadgetInfo.d.ts";
import type { BossChestInfo } from "./BossChestInfo.d.ts";
import type { BlossomChestInfo } from "./BlossomChestInfo.d.ts";
import type { MpPlayRewardInfo } from "./MpPlayRewardInfo.d.ts";
import type { GadgetGeneralRewardInfo } from "./GadgetGeneralRewardInfo.d.ts";
import type { OfferingInfo } from "./OfferingInfo.d.ts";
import type { FoundationInfo } from "./FoundationInfo.d.ts";
import type { VehicleInfo } from "./VehicleInfo.d.ts";
import type { EchoShellInfo } from "./EchoShellInfo.d.ts";
import type { ScreenInfo } from "./ScreenInfo.d.ts";
import type { FishPoolInfo } from "./FishPoolInfo.d.ts";
import type { CustomGadgetTreeInfo } from "./CustomGadgetTreeInfo.d.ts";
import type { RoguelikeGadgetInfo } from "./RoguelikeGadgetInfo.d.ts";
import type { Unk2800FHMOLALLEEN } from "./Unk2800_FHMOLALLEEN.d.ts";

export interface SceneGadgetInfo {
  GadgetId?: number;
  GroupId?: number;
  ConfigId?: number;
  OwnerEntityId?: number;
  BornType?: GadgetBornType;
  GadgetState?: number;
  GadgetType?: number;
  IsShowCutscene?: boolean;
  AuthorityPeerId?: number;
  IsEnableInteract?: boolean;
  InteractId?: number;
  MarkFlag?: number;
  PropOwnerEntityId?: number;
  Platform?: PlatformInfo;
  InteractUidList?: number[];
  DraftId?: number;
  GadgetTalkState?: number;
  PlayInfo?: GadgetPlayInfo;
  TrifleItem?: Item | undefined;
  GatherGadget?: GatherGadgetInfo | undefined;
  Worktop?: WorktopInfo | undefined;
  ClientGadget?: ClientGadgetInfo | undefined;
  Weather?: WeatherInfo | undefined;
  AbilityGadget?: AbilityGadgetInfo | undefined;
  StatueGadget?: StatueGadgetInfo | undefined;
  BossChest?: BossChestInfo | undefined;
  BlossomChest?: BlossomChestInfo | undefined;
  MpPlayReward?: MpPlayRewardInfo | undefined;
  GeneralReward?: GadgetGeneralRewardInfo | undefined;
  OfferingInfo?: OfferingInfo | undefined;
  FoundationInfo?: FoundationInfo | undefined;
  VehicleInfo?: VehicleInfo | undefined;
  ShellInfo?: EchoShellInfo | undefined;
  ScreenInfo?: ScreenInfo | undefined;
  FishPoolInfo?: FishPoolInfo | undefined;
  CustomGadgetTreeInfo?: CustomGadgetTreeInfo | undefined;
  RoguelikeGadgetInfo?: RoguelikeGadgetInfo | undefined;
  NightCrowGadgetInfo?: Unk2800FHMOLALLEEN | undefined;
}